pyproject.toml
[project]
name = "compitum"
version = "0.1.0"
description = "compitum: A Production-Ready, Geometrically-Aware AI Router"
authors = [{ name="compitum authors", email="dev@compitum.org" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
 "numpy>=1.24",
 "scipy>=1.10",
 "scikit-learn>=1.3",
 "pydantic>=2.7",
 "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = [
 "pytest>=8.0",
 "hypothesis>=6.98",
 "pytest-cov>=5.0",
 "ruff>=0.5.0",
 "mypy>=1.10",
 "lightgbm>=4.3 ; python_version < '3.13'",
 "types-PyYAML",
 "cosmic-ray",
 "bandit"
]

[project.scripts]
compitum = "compitum.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
compitum = ["py.typed"]

# --------------------------- 
# Ruff
# --------------------------- 
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

# --------------------------- 
# MyPy
# --------------------------- 
[tool.mypy]
files = ["compitum", "tests"]
disallow_untyped_defs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["numpy.*", "scipy.*", "sklearn.*", "yaml.*", "lightgbm.*"]
ignore_missing_imports = true

# --------------------------- 
# PyTest
# --------------------------- 
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
  "-ra",
  "--cov=compitum",
  "--cov-branch",
  "--cov-report=term-missing",
  "-q"
]
markers = ["invariants: tests for core system invariants"]

# --------------------------- 
# Coverage
# --------------------------- 
[tool.coverage.run]
branch = true
source = ["compitum"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
